mlops_deployment/requirements.txt
# Core ML libraries
scikit-learn>=1.5.0
pandas>=2.0.0
numpy>=1.26.0
matplotlib>=3.7.0
seaborn>=0.12.0
plotly>=5.17.0
kaleido>=0.2.1
shap>=0.42.0
reportlab>=4.0.0
fpdf2>=2.7.0
streamlit-option-menu>=0.3.6
streamlit-extras>=0.3.0

# API framework
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
pydantic>=2.9.0

# Web UI
streamlit>=1.40.0
requests>=2.32.0

# Utilities
python-dotenv>=1.0.0
pyyaml>=6.0.2
joblib>=1.4.0

# Optional: for development
pytest>=8.0.0
black>=24.0.0

